+++ src/base/b-virtual-scroll/demo.js
@@ -32,7 +32,8 @@ const suits = {
 	demo: [{
 		attrs: {
 			...baseAttrs,
-			':dataProvider': s('demo.Pagination')
+			':dataProvider': s('demo.Pagination'),
+			':request': '{get: {id: "b-virtual:init-load", chunkSize: 10}}'
 		},
 
 		content: {
@@ -98,20 +99,30 @@ const suits = {
 			...baseContent,
 			empty: slots.empty
 		}
+	}],
+
+	doubleComponents: [{
+		attrs: {
+			...baseAttrs,
+			':dataProvider': s('demo.Pagination'),
+			':request': '{get: {chunkSize: 10}}'
+		},
+
+		content: {
+			...baseContent
+		}
 	}, {
 		attrs: {
 			...baseAttrs,
 			':dataProvider': s('demo.Pagination'),
-			':dbConverter': '({data}) => ({data: data.splice(0, 4)})',
-			':request': '{get: {chunkSize: 8, total: 8, id: "b-virtual:slot-empty"}}',
+			':request': '{get: {id: "b-virtual:init-load", chunkSize: 10}}',
 			'id': 'second'
 		},
 
 		content: {
-			...baseContent,
-			empty: slots.empty
+			...baseContent
 		}
-	}]
+	}],
 };
 
 module.exports = suits;