+++ src/base/b-virtual-scroll/test/runners/render-truncated.js
@@ -18,9 +18,9 @@ module.exports = async (page, {componentSelector, component: c}) => {
 	describe('b-virtual-scroll', () => {
 		it('renders truncated data chunks to the page', async () => {
 			const [chunkSize, total, requestChunkSize, convertedLength] = await Promise.all([
-				h.getField(c, 'chunkSize'),
-				h.getField(c, 'request.get.total'),
-				h.getField(c, 'request.get.chunkSize'),
+				c.evaluate((ctx) => ctx.chunkSize),
+				c.evaluate((ctx) => ctx.request.get.total),
+				c.evaluate((ctx) => ctx.request.get.chunkSize),
 				c.evaluate((ctx) => ctx.dbConverter({data: ctx.global.Array(100)}).data.length)
 			]);
 
@@ -28,25 +28,25 @@ module.exports = async (page, {componentSelector, component: c}) => {
 				hasSkeletons = await c.evaluate((ctx) => Boolean(ctx.vdom.getSlot('tombstone'))),
 				totalGivenDataToRender = total / (requestChunkSize / convertedLength);
 
-			await h.waitItemsCountGreaterThan(page, 0, componentSelector);
+			await h.waitItemsCountGreaterThan(page, 0, {componentSelector});
 
 			expect(await c.evaluate((ctx) => ctx.$refs.container.childElementCount)).toBe(chunkSize);
 
 			await h.scrollToPageBottom(page);
 
 			if (hasSkeletons) {
-				await h.waitForRefDisplay(page, componentSelector, 'tombstones', '');
+				await h.waitForRefDisplay(page, {componentSelector}, 'tombstones', '');
 				expect(await c.evaluate((ctx) => ctx.$refs.tombstones.style.display)).toBe('');
 			}
 
-			await h.waitItemsCountGreaterThan(page, chunkSize, componentSelector);
+			await h.waitItemsCountGreaterThan(page, chunkSize, {componentSelector});
 
 			expect(await c.evaluate((ctx) => ctx.$refs.container.childElementCount)).toBe(totalGivenDataToRender);
 			expect(await c.evaluate((ctx) => ctx.chunkRequest.isDone)).toBe(true);
 			expect(await c.evaluate((ctx) => ctx.chunkRequest.pendingData.length)).toBe(0);
 
 			if (hasSkeletons) {
-				await h.waitForRefDisplay(page, componentSelector, 'tombstones', 'none');
+				await h.waitForRefDisplay(page, {componentSelector}, 'tombstones', 'none');
 				expect(await c.evaluate((ctx) => ctx.$refs.tombstones.style.display)).toBe('none');
 			}
 		});