+++ tests/helpers.js
@@ -7,34 +7,6 @@
  */
 
 /**
- * Returns a property from an object by the specified path
- *
- * @param {?} object
- * @param {string} path
- */
-module.exports.getField = async function (object, path) {
-	if (!object) {
-		return undefined;
-	}
-
-	const
-		chunks = path.split('.');
-
-	let
-		res = object;
-
-	for (let i = 0; i < chunks.length; i++) {
-		if (res == null) {
-			return undefined;
-		}
-
-		res = await res.getProperty(chunks[i]);
-	}
-
-	return await res.jsonValue();
-};
-
-/**
  * Returns a component context by the specified selector
  *
  * @param {?} page